requirements.txt
pandas
seaborn
os
matplotlib.plyplot
faker
random
os
scikit-learn
python-dotenv
sqlalchemy
pathlib